requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.6
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
pinecone-client==2.2.4
